backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
pydantic-settings==2.10.1
python-dotenv==1.1.0
google-generativeai==0.8.4
SQLAlchemy==2.0.36
pymysql==1.1.1
passlib[bcrypt]==1.7.4
python-jose==3.3.0
cryptography==44.0.0
bcrypt==3.2.2
